final/requirements.txt
django
djangorestframework
markdown
django-filter
pyjwt
Pillow

# swagger
django-rest-swagger
drf-yasg
